ALA-GUI/requirements.txt
# ALA-GUI Requirements
# Python 3.9+ (Florence-2 only)
# Python 3.10+ (Florence-2 + SAM2 segmentation)

# PyQt6 - GUI Framework
PyQt6==6.6.1
PyQt6-Qt6==6.6.1
PyQt6-sip==13.6.0
PyQt6-WebEngine==6.6.0

# Testing
pytest==7.4.3
pytest-qt==4.3.1
pytest-cov==4.1.0
pytest-mock==3.12.0

# Image Processing
opencv-python==4.9.0.80
Pillow==10.2.0
numpy==1.24.3

# Machine Learning Models (from existing ALA)
torch==2.1.2
torchvision==0.16.2
transformers==4.36.2

# Florence-2 dependencies (flash_attn optional)
timm>=0.9.0
einops>=0.7.0

# Web Interface
gradio==4.14.0
fastapi==0.109.0
uvicorn==0.27.0

# Data Formats
pyyaml==6.0.1

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1

# Code Quality
black==24.1.1
flake8==7.0.0
mypy==1.8.0
pylint==3.0.3
isort==5.13.2
bandit==1.7.6

# Model packages (existing ALA dependencies)
ultralytics==8.1.11
supervision==0.18.0
